shift metrics into child datastore + udpate helm/ui docs
