--- charts/filament/templates/metrics-configmap.yaml
@@ -1,13 +1,0 @@
-{{- if .Values.metrics.enabled }}
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: {{ include "filament.metrics.fullname" . }}
-  namespace: {{ include "filament.namespace" . }}
-data:
-  # Metrics store — reads the same runs table server/control-plane persist to.
-  METRICSSTORE_PROVIDER: {{ .Values.persistence.type | quote }}
-
-  # Serving
-  SERVER_ADDR: ":{{ .Values.metrics.service.port }}"
-{{- end }}
