--- charts/filament/templates/metrics-deployment.yaml
@@ -1,62 +1,0 @@
-{{- if .Values.metrics.enabled }}
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: {{ include "filament.metrics.fullname" . }}
-  namespace: {{ include "filament.namespace" . }}
-  labels:
-    {{- include "filament.metrics.labels" . | nindent 4 }}
-spec:
-  {{- if not .Values.metrics.autoscaling.enabled }}
-  replicas: {{ .Values.metrics.replicas }}
-  {{- end }}
-  selector:
-    matchLabels:
-      {{- include "filament.metrics.selectorLabels" . | nindent 6 }}
-  template:
-    metadata:
-      annotations:
-        checksum/config: {{ include (print $.Template.BasePath "/metrics-configmap.yaml") . | sha256sum }}
-        checksum/secret: {{ include (print $.Template.BasePath "/secret.yaml") . | sha256sum }}
-      labels:
-        {{- include "filament.metrics.selectorLabels" . | nindent 8 }}
-    spec:
-      serviceAccountName: {{ include "filament.metrics.serviceAccountName" . }}
-      {{- with .Values.metrics.image.pullSecrets }}
-      imagePullSecrets:
-        {{- toYaml . | nindent 8 }}
-      {{- end }}
-      containers:
-        - name: metrics
-          image: {{ .Values.metrics.image.repository }}:{{ .Values.metrics.image.tag | default .Chart.AppVersion }}
-          imagePullPolicy: {{ .Values.metrics.image.pullPolicy }}
-          ports:
-            - name: http
-              containerPort: {{ .Values.metrics.service.port }}
-              protocol: TCP
-          startupProbe:
-            httpGet:
-              path: /livez
-              port: http
-            periodSeconds: 5
-            failureThreshold: 30
-          readinessProbe:
-            httpGet:
-              path: /readyz
-              port: http
-            periodSeconds: 10
-            timeoutSeconds: 5
-          livenessProbe:
-            httpGet:
-              path: /livez
-              port: http
-            periodSeconds: 20
-            timeoutSeconds: 5
-          envFrom:
-            - configMapRef:
-                name: {{ include "filament.metrics.fullname" . }}
-            - secretRef:
-                name: {{ include "filament.secretName" . }}
-          resources:
-            {{- toYaml .Values.metrics.resources | nindent 12 }}
-{{- end }}
