--- charts/filament/templates/metrics-hpa.yaml
@@ -1,40 +1,0 @@
-{{- if and .Values.metrics.enabled .Values.metrics.autoscaling.enabled }}
-apiVersion: autoscaling/v2
-kind: HorizontalPodAutoscaler
-metadata:
-  name: {{ include "filament.metrics.fullname" . }}
-  namespace: {{ include "filament.namespace" . }}
-  labels:
-    {{- include "filament.metrics.labels" . | nindent 4 }}
-spec:
-  minReplicas: {{ .Values.metrics.autoscaling.minReplicas }}
-  maxReplicas: {{ .Values.metrics.autoscaling.maxReplicas }}
-  scaleTargetRef:
-    apiVersion: apps/v1
-    kind: Deployment
-    name: {{ include "filament.metrics.fullname" . }}
-  metrics:
-    - type: Resource
-      resource:
-        name: cpu
-        target:
-          type: Utilization
-          averageUtilization: {{ .Values.metrics.autoscaling.targetCpu }}
-  behavior:
-    scaleDown:
-      stabilizationWindowSeconds: 300
-      policies:
-        - type: Percent
-          value: 10
-          periodSeconds: 60
-    scaleUp:
-      stabilizationWindowSeconds: 0
-      policies:
-        - type: Percent
-          value: 100
-          periodSeconds: 15
-        - type: Pods
-          value: 4
-          periodSeconds: 15
-      selectPolicy: Max
-{{- end }}
