--- charts/filament/templates/metrics-service.yaml
@@ -1,18 +1,0 @@
-{{- if .Values.metrics.enabled }}
-apiVersion: v1
-kind: Service
-metadata:
-  name: {{ include "filament.metrics.fullname" . }}
-  namespace: {{ include "filament.namespace" . }}
-  labels:
-    {{- include "filament.metrics.labels" . | nindent 4 }}
-spec:
-  type: ClusterIP
-  ports:
-    - name: http
-      port: {{ .Values.metrics.service.port }}
-      targetPort: http
-      protocol: TCP
-  selector:
-    {{- include "filament.metrics.selectorLabels" . | nindent 4 }}
-{{- end }}
