--- charts/filament/templates/metrics-serviceaccount.yaml
@@ -1,13 +1,0 @@
-{{- if and .Values.metrics.enabled (not .Values.metrics.serviceAccount.name) }}
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: {{ include "filament.metrics.serviceAccountName" . }}
-  namespace: {{ include "filament.namespace" . }}
-  labels:
-    {{- include "filament.metrics.labels" . | nindent 4 }}
-  {{- with .Values.metrics.serviceAccount.annotations }}
-  annotations:
-    {{- toYaml . | nindent 4 }}
-  {{- end }}
-{{- end }}
