--- cmd/metrics/main.go
@@ -1,73 +1,0 @@
-// Command metrics runs MetricsService: it answers run metrics queries by
-// reading the same runs table server/control-plane persist to.
-package main
-
-import (
-	"context"
-	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"os"
-	"os/signal"
-	"syscall"
-	"time"
-
-	"github.com/galaxy-io/filament/cmd/internal/metricsstore"
-	"github.com/galaxy-io/filament/metrics"
-)
-
-func main() {
-	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
-	err := run(ctx)
-	stop()
-	if err != nil {
-		log.Fatal(err)
-	}
-}
-
-func run(ctx context.Context) error {
-	store, err := metricsstore.FromEnv(ctx)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		if c, ok := store.(io.Closer); ok {
-			_ = c.Close()
-		}
-	}()
-
-	addr := os.Getenv("SERVER_ADDR")
-	if addr == "" {
-		addr = ":8081"
-	}
-	mux := http.NewServeMux()
-	mux.HandleFunc("/livez", func(w http.ResponseWriter, _ *http.Request) {
-		w.WriteHeader(http.StatusOK)
-	})
-	mux.HandleFunc("/readyz", func(w http.ResponseWriter, r *http.Request) {
-		ctx, cancel := context.WithTimeout(r.Context(), 2*time.Second)
-		defer cancel()
-		if err := store.Ping(ctx); err != nil {
-			http.Error(w, "not ready: "+err.Error(), http.StatusServiceUnavailable)
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-	})
-
-	metrics.New(store).Mount(mux)
-
-	srv := &http.Server{Addr: addr, Handler: mux, ReadHeaderTimeout: 10 * time.Second}
-	errCh := make(chan error, 1)
-	go func() { errCh <- srv.ListenAndServe() }()
-	fmt.Println("metrics:", "http://localhost"+addr)
-
-	select {
-	case err := <-errCh:
-		return err
-	case <-ctx.Done():
-		shutCtx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-		defer cancel()
-		return srv.Shutdown(shutCtx)
-	}
-}
