+++ datastore/postgres/metrics/templates/aggregate.sql.tmpl
